fix(job): consolidate Kueue lifecycle and queue management into kueue.go (#6167)
