+++ pkg/orchestrator/gke/infra_manager.go
@@ -20,14 +20,10 @@ import (
 	"encoding/json"
 	"fmt"
 	"hpc-toolkit/pkg/logging"
-	"hpc-toolkit/pkg/shell"
 	"io"
 	"net/http"
 	"os"
 	"path/filepath"
-	"slices"
-	"sort"
-	"strconv"
 	"strings"
 	"time"
 
@@ -39,10 +35,6 @@ import (
 //go:embed templates/*
 var templatesFS embed.FS
 
-// defaultKueueVersion is the fallback version of Kueue to install.
-// ATTENTION: For cluster creation the corresponding default version
-// is in modules/management/kubectl-apply/variables.tf.
-const defaultKueueVersion = "v0.15.2"
 const defaultJobSetVersion = "v0.10.1"
 
 func (g *GKEOrchestrator) checkAndInstallJobSetCRD() error {
@@ -50,9 +42,17 @@ func (g *GKEOrchestrator) checkAndInstallJobSetCRD() error {
 		return err
 	} else if installed {
 		logging.Info("JobSet CRD found. Verifying Webhook health...")
-		cmdEndpoints := g.executor.ExecuteCommand("kubectl", "get", "endpoints", "jobset-webhook-service", "-n", "jobset-system", "-o", "jsonpath={.subsets[*].addresses[*].ip}")
-		if cmdEndpoints.ExitCode == 0 && strings.TrimSpace(cmdEndpoints.Stdout) != "" {
-			logging.Info("JobSet Webhook is healthy.")
+		cmdEndpoints := g.executor.ExecuteCommand("kubectl", "get", "endpointslice", "-l", "kubernetes.io/service-name=jobset-webhook-service", "-n", "jobset-system", "-o", "json")
+		if cmdEndpoints.ExitCode == 0 {
+			var eps k8sEndpointSliceList
+			if err := json.Unmarshal([]byte(cmdEndpoints.Stdout), &eps); err != nil {
+				logging.Warn("Failed to parse JobSet endpointslice JSON: %v", err)
+			} else if eps.HasReadyEndpoint() {
+				logging.Info("JobSet Webhook is healthy.")
+				return nil
+			}
+		} else if strings.Contains(strings.ToLower(cmdEndpoints.Stderr), "forbidden") {
+			logging.Warn("Insufficient RBAC permissions to read JobSet webhook endpoints (403 Forbidden). Assuming JobSet is healthy in shared cluster.")
 			return nil
 		}
 		logging.Info("JobSet Webhook endpoints not found. Proceeding with re-installation/fix...")
@@ -60,539 +60,6 @@ func (g *GKEOrchestrator) checkAndInstallJobSetCRD() error {
 
 	jobSetManifestsURL := fmt.Sprintf("https://github.com/kubernetes-sigs/jobset/releases/download/%s/manifests.yaml", defaultJobSetVersion)
 	return g.installJobSetCRD(jobSetManifestsURL)
-}
-
-func (g *GKEOrchestrator) CheckAndInstallKueue(version string, clusterName string, clusterLocation string) error {
-	kueueCRDInstalled, _ := g.isKueueInstalled()
-	kueueDeploymentInstalled, _ := g.isKueueDeploymentInstalled()
-	currentVersion, _ := g.GetKueueVersion()
-
-	if version == "" {
-		version = defaultKueueVersion
-	}
-
-	var reinstallReason string
-	needReinstall := false
-
-	// 1. Minimum version check for KUEUE
-	if currentVersion != "" && g.isVersionBelow(currentVersion, version) {
-		needReinstall = true
-		reinstallReason = fmt.Sprintf("Current Kueue version %s is below target %s.", currentVersion, version)
-	}
-
-	// 2. Check if basic installation is missing
-	if !needReinstall && (!kueueCRDInstalled || !kueueDeploymentInstalled) {
-		needReinstall = true
-		reinstallReason = "Kueue installation is incomplete (CRD or Deployment missing)."
-	}
-
-	// 3. Check if webhook is healthy (only if not already deciding to reinstall)
-	if !needReinstall && g.waitForKueueWebhook() != nil {
-		needReinstall = true
-		reinstallReason = "Kueue webhook health check failed. Treating as broken."
-	}
-
-	if needReinstall {
-		if g.checkDynamicSlicingViaGKE() {
-			return fmt.Errorf("automatic Kueue installation blocked: we detected that cluster %s is set up for Dynamic-slicing (found 'PROVISION_ONLY' in the node pool's placementPolicy). Wiping Kueue would corrupt your custom topology configurations. Please install Kueue and the required custom CRDs manually", clusterName)
-		}
-
-		if err := g.checkKueueInstallPermissions(version); err != nil {
-			return err
-		}
-
-		if err := g.handleKueueReinstallation(version, reinstallReason); err != nil {
-			return err
-		}
-	}
-
-	logging.Info("Kueue is already installed.")
-	return nil
-}
-
-func (g *GKEOrchestrator) ensurePriorityClassesInstalled() error {
-	hasUserClasses, err := g.hasUserPriorityClasses()
-	if err != nil {
-		return err
-	}
-
-	if hasUserClasses {
-		return nil // Skip installation entirely
-	}
-
-	logging.Info("No user-defined PriorityClasses found. Installing defaults...")
-	return g.installPriorityClasses()
-}
-
-func (g *GKEOrchestrator) handleKueueReinstallation(targetVersion string, reason string) error {
-	promptMsg := fmt.Sprintf("%s\nKueue requires re-installation using %s.\nWARNING: This deletes all queued and suspended workloads in this cluster before proceeding.\nReplying 'no' will cause an immediate exit and you will have to do the re-installation manually. Proceed?", reason, targetVersion)
-	if !shell.PromptYesNo(promptMsg) {
-		return fmt.Errorf("user declined to re-install Kueue. Exiting.")
-	}
-
-	logging.Info("Proceeding with clean re-installation of Kueue...")
-	if err := g.DeleteAllKueueResources(); err != nil {
-		return fmt.Errorf("failed to delete Kueue resources: %w", err)
-	}
-
-	return g.installKueue(targetVersion)
-}
-
-func (g *GKEOrchestrator) isVersionBelow(current, target string) bool {
-	curMajor, curMinor, curPatch := parseVersion(current)
-	defMajor, defMinor, defPatch := parseVersion(target)
-
-	return curMajor < defMajor || (curMajor == defMajor && curMinor < defMinor) || (curMajor == defMajor && curMinor == defMinor && curPatch < defPatch)
-}
-
-func (g *GKEOrchestrator) DeleteKueueDeployment() error {
-	logging.Info("Deleting Kueue deployment...")
-	res := g.executor.ExecuteCommand("kubectl", "delete", "deployment", "kueue-controller-manager", "-n", "kueue-system", "--ignore-not-found")
-	if res.ExitCode != 0 {
-		return fmt.Errorf("failed to delete Kueue deployment: %s\n%s", res.Stderr, res.Stdout)
-	}
-	return nil
-}
-
-func (g *GKEOrchestrator) DeleteAllKueueResources() error {
-	crds := []string{
-		"admissionchecks.kueue.x-k8s.io",
-		"clusterqueues.kueue.x-k8s.io",
-		"cohorts.kueue.x-k8s.io",
-		"localqueues.kueue.x-k8s.io",
-		"multikueueclusters.kueue.x-k8s.io",
-		"multikueueconfigs.kueue.x-k8s.io",
-		"provisioningrequestconfigs.kueue.x-k8s.io",
-		"resourceflavors.kueue.x-k8s.io",
-		"topologies.kueue.x-k8s.io",
-		"workloadpriorityclasses.kueue.x-k8s.io",
-		"workloads.kueue.x-k8s.io",
-	}
-
-	resourceList := strings.Join(crds, ",")
-	logging.Info("Deleting all Kueue resources...")
-	g.executor.ExecuteCommand("kubectl", "delete", resourceList, "--all", "--ignore-not-found", "--wait=false")
-
-	logging.Info("Deleting Kueue CRDs...")
-	args := append([]string{"delete", "crd", "--ignore-not-found"}, crds...)
-	res := g.executor.ExecuteCommand("kubectl", args...)
-	if res.ExitCode != 0 {
-		return fmt.Errorf("failed to delete Kueue CRDs: %s\n%s", res.Stderr, res.Stdout)
-	}
-
-	return g.DeleteKueueDeployment()
-}
-
-func (g *GKEOrchestrator) isKueueInstalled() (bool, error) {
-	res := g.executor.ExecuteCommand("kubectl", "get", "crd", "clusterqueues.kueue.x-k8s.io")
-	if res.ExitCode == 0 {
-		logging.Info("Kueue CRD found.")
-		return true, nil
-	}
-	if strings.Contains(res.Stderr, "not found") || strings.Contains(res.Stdout, "NotFound") {
-		logging.Info("Kueue CRD not found.")
-		return false, nil
-	}
-	return false, fmt.Errorf("failed to check for Kueue CRD: %s\n%s", res.Stderr, res.Stdout)
-}
-
-func (g *GKEOrchestrator) isKueueDeploymentInstalled() (bool, error) {
-	res := g.executor.ExecuteCommand("kubectl", "get", "deployment", "kueue-controller-manager", "-n", "kueue-system")
-	if res.ExitCode == 0 {
-		logging.Info("Kueue deployment found.")
-		return true, nil
-	}
-	if strings.Contains(res.Stderr, "not found") || strings.Contains(res.Stdout, "NotFound") {
-		logging.Info("Kueue deployment not found.")
-		return false, nil
-	}
-	return false, fmt.Errorf("failed to check for Kueue deployment: %s\n%s", res.Stderr, res.Stdout)
-}
-
-func (g *GKEOrchestrator) GetKueueVersion() (string, error) {
-	res := g.executor.ExecuteCommand("kubectl", "get", "deployment", "kueue-controller-manager", "-n", "kueue-system", "-o", "jsonpath={.spec.template.spec.containers[0].image}")
-	if res.ExitCode != 0 {
-		return "", fmt.Errorf("failed to get Kueue version: %s\n%s", res.Stderr, res.Stdout)
-	}
-	image := strings.TrimSpace(res.Stdout)
-	// The image string might be "registry.k8s.io/kueue/kueue:v0.6.3"
-	// We want to extract "v0.6.3"
-	parts := strings.Split(image, ":")
-	if len(parts) < 2 {
-		return "", fmt.Errorf("unexpected image format for Kueue: %s", image)
-	}
-	version := parts[len(parts)-1]
-	return version, nil
-}
-
-func (g *GKEOrchestrator) getClusterPriorityClasses() ([]string, error) {
-	res := g.executor.ExecuteCommand("kubectl", "get", "priorityclass", "-o", "jsonpath={.items[*].metadata.name}")
-	if res.ExitCode != 0 {
-		return nil, fmt.Errorf("failed to list priority classes: %s", res.Stderr)
-	}
-	return strings.Fields(res.Stdout), nil
-}
-
-func (g *GKEOrchestrator) hasUserPriorityClasses() (bool, error) {
-	existing, err := g.getClusterPriorityClasses()
-	if err != nil {
-		return false, err
-	}
-
-	for _, name := range existing {
-		if strings.HasPrefix(name, "system-") || strings.HasPrefix(name, "gke-") {
-			continue
-		}
-		logging.Info("Pre-existing PriorityClass '%s' found. Skipping default PriorityClass installation.", name)
-		return true, nil
-	}
-	return false, nil
-}
-
-func (g *GKEOrchestrator) installKueue(version string) error {
-	logging.Info("Installing Kueue version %s...", version)
-	kueueManifestsURL := fmt.Sprintf("https://github.com/kubernetes-sigs/kueue/releases/download/%s/manifests.yaml", version)
-	manifestBytes, err := g.downloadManifests(kueueManifestsURL)
-	if err != nil {
-		return err
-	}
-
-	cleanedManifests, err := g.cleanAndProcessManifests(manifestBytes, nil)
-	if err != nil {
-		return err
-	}
-
-	if err := g.applyManifests(cleanedManifests, "kueue.yaml"); err != nil {
-		return err
-	}
-
-	logging.Info("Kueue components applied successfully.")
-
-	if err := g.waitForKueueWebhook(); err != nil {
-		return err
-	}
-
-	return g.installKueueResources(defaultClusterQueue, defaultLocalQueue)
-}
-
-func (g *GKEOrchestrator) installPriorityClasses() error {
-	logging.Info("Installing Kueue PriorityClasses...")
-	tmpl, err := g.parseGKETemplate("priority_classes.tmpl")
-	if err != nil {
-		return err
-	}
-
-	var buf bytes.Buffer
-	if err := tmpl.Execute(&buf, nil); err != nil {
-		return fmt.Errorf("failed to execute priority_classes template: %w", err)
-	}
-
-	logging.Info("Applying Kueue priority classes...")
-	if err := g.applyManifests(buf.Bytes(), "kueue_priority_classes.yaml"); err != nil {
-		return fmt.Errorf("failed to apply Kueue priority classes: %w", err)
-	}
-
-	return nil
-}
-
-func (g *GKEOrchestrator) installKueueResources(cqName string, lqName string) error {
-	logging.Info("Installing Kueue resources (ClusterQueue, LocalQueue)...")
-
-	hasUserClasses, err := g.hasUserPriorityClasses()
-	if err != nil {
-		return err
-	}
-	if !hasUserClasses {
-		if err := g.installPriorityClasses(); err != nil {
-			return err
-		}
-	}
-
-	// Install ClusterQueue
-	clusterQueueBytes, err := g.renderClusterQueue(cqName)
-	if err != nil {
-		return err
-	}
-	if err := g.applyManifests(clusterQueueBytes, "cluster-queue.yaml"); err != nil {
-		return err
-	}
-
-	// Install LocalQueue
-	localQueueTmpl, err := g.parseGKETemplate("local_queue.tmpl")
-	if err != nil {
-		return fmt.Errorf("failed to parse local_queue.tmpl: %w", err)
-	}
-	var localQueueBuf bytes.Buffer
-	if err := localQueueTmpl.Execute(&localQueueBuf, struct {
-		Namespace        string
-		LocalQueueName   string
-		ClusterQueueName string
-	}{"default", lqName, cqName}); err != nil {
-		return fmt.Errorf("failed to execute local_queue.tmpl template: %w", err)
-	}
-	if err := g.applyManifests(localQueueBuf.Bytes(), "local-queue.yaml"); err != nil {
-		return err
-	}
-
-	logging.Info("Kueue resources installed successfully.")
-	return nil
-}
-
-func (g *GKEOrchestrator) isResourceActive(staticAmount int, napLimitKey string) bool {
-	return staticAmount > 0 || (g.napEnabled && g.napLimits[napLimitKey] > 0)
-}
-
-func (g *GKEOrchestrator) isAccelActive(staticAmount int, napLimitKey string, fname string, flavorSub string) bool {
-	if staticAmount > 0 {
-		return true
-	}
-	return g.napEnabled && g.napLimits[napLimitKey] > 0 && strings.Contains(strings.ToLower(fname), flavorSub)
-}
-
-func (g *GKEOrchestrator) calculateCoveredResources() (map[string]bool, map[string]bool) {
-	mainMap := make(map[string]bool)
-	pathwaysMap := make(map[string]bool)
-
-	for fname, fc := range g.capacity.Flavors {
-		if fname == "pathways-flavor" {
-			if g.isResourceActive(fc.CPUs, "cpu") {
-				pathwaysMap["cpu"] = true
-			}
-			if g.isResourceActive(fc.MemoryGi, "memory") {
-				pathwaysMap["memory"] = true
-			}
-		} else {
-			if g.isResourceActive(fc.CPUs, "cpu") {
-				mainMap["cpu"] = true
-			}
-			if g.isResourceActive(fc.MemoryGi, "memory") {
-				mainMap["memory"] = true
-			}
-			if g.isAccelActive(fc.GPUs, "nvidia.com/gpu", fname, "nvidia") {
-				mainMap["nvidia.com/gpu"] = true
-			}
-			if g.isAccelActive(fc.TPUs, "google.com/tpu", fname, "tpu") {
-				mainMap["google.com/tpu"] = true
-			}
-		}
-	}
-	return mainMap, pathwaysMap
-}
-
-func isTPUFlavor(fname string) bool {
-	lower := strings.ToLower(fname)
-	return strings.Contains(lower, "tpu") || strings.Contains(lower, "ct")
-}
-
-func isGPUFlavor(fname string) bool {
-	return strings.Contains(strings.ToLower(fname), "nvidia")
-}
-
-func (g *GKEOrchestrator) getNAPNominalQuota(resName string, fname string) (interface{}, bool) {
-	if !g.napEnabled {
-		return nil, false
-	}
-	lookupKey := resName
-	if resName == "nvidia.com/gpu" || resName == "google.com/tpu" {
-		specificKey := strings.TrimPrefix(fname, "flavor-")
-		if _, ok := g.napLimits[specificKey]; ok {
-			lookupKey = specificKey
-		}
-	}
-	limit, ok := g.napLimits[lookupKey]
-	if !ok {
-		return nil, false
-	}
-	// Only apply TPU/GPU limit to matching flavors
-	if resName == "google.com/tpu" && !isTPUFlavor(fname) {
-		return 0, true
-	}
-	if resName == "nvidia.com/gpu" && !isGPUFlavor(fname) {
-		return 0, true
-	}
-	if resName == "memory" {
-		return fmt.Sprintf("%dG", limit), true // limit is in GB in GKE describe
-	}
-	return limit, true
-}
-
-func getStaticNominalQuota(resName string, fc FlavorCapacity) interface{} {
-	switch resName {
-	case "cpu":
-		return fc.CPUs
-	case "memory":
-		return fmt.Sprintf("%dGi", fc.MemoryGi)
-	case "nvidia.com/gpu":
-		return fc.GPUs
-	case "google.com/tpu":
-		return fc.TPUs
-	default:
-		return 0
-	}
-}
-
-func (g *GKEOrchestrator) getNominalQuota(resName string, fc FlavorCapacity, fname string) interface{} {
-	if val, ok := g.getNAPNominalQuota(resName, fname); ok {
-		return val
-	}
-	return getStaticNominalQuota(resName, fc)
-}
-
-func (g *GKEOrchestrator) renderClusterQueue(name string) ([]byte, error) {
-	mainCoveredResourcesMap, pathwaysCoveredResourcesMap := g.calculateCoveredResources()
-
-	_, hasPathways := g.capacity.Flavors["pathways-flavor"]
-
-	// If pathways is active, we unify the resource groups to prevent Kueue from
-	// merging incompatible node selectors for TPU worker pods (which request both
-	// TPU and CPU/Memory).
-	if hasPathways {
-		for res := range pathwaysCoveredResourcesMap {
-			mainCoveredResourcesMap[res] = true
-		}
-		pathwaysCoveredResourcesMap = make(map[string]bool) // Clear to prevent second group generation
-	}
-
-	// Pass 2: Build flavor lists with matched resources
-	var mainFlavors []map[string]interface{}
-	var pathwaysFlavors []map[string]interface{}
-
-	var fnames []string
-	for fname := range g.capacity.Flavors {
-		fnames = append(fnames, fname)
-	}
-	sort.Strings(fnames)
-
-	for _, fname := range fnames {
-		fc := g.capacity.Flavors[fname]
-		isPathways := (fname == "pathways-flavor")
-
-		coveredMap := mainCoveredResourcesMap
-		if isPathways && !hasPathways {
-			coveredMap = pathwaysCoveredResourcesMap
-		}
-
-		resources := g.buildFlavorResources(coveredMap, fc, fname)
-		if len(resources) > 0 {
-			flavor := map[string]interface{}{
-				"name":      fname,
-				"resources": resources,
-			}
-			if isPathways && !hasPathways {
-				pathwaysFlavors = append(pathwaysFlavors, flavor)
-			} else {
-				mainFlavors = append(mainFlavors, flavor)
-			}
-		}
-	}
-
-	var resourceGroups []map[string]interface{}
-	if rg := buildResourceGroup(mainCoveredResourcesMap, mainFlavors); rg != nil {
-		resourceGroups = append(resourceGroups, rg)
-	}
-	if rg := buildResourceGroup(pathwaysCoveredResourcesMap, pathwaysFlavors); rg != nil {
-		resourceGroups = append(resourceGroups, rg)
-	}
-
-	cqMap := map[string]interface{}{
-		"apiVersion": "kueue.x-k8s.io/" + kueueAPIVersion,
-		"kind":       "ClusterQueue",
-		"metadata": map[string]interface{}{
-			"name": name,
-		},
-		"spec": map[string]interface{}{
-			"namespaceSelector": map[string]interface{}{},
-			"queueingStrategy":  "BestEffortFIFO",
-			"resourceGroups":    resourceGroups,
-		},
-	}
-
-	cqBytes, err := yaml.Marshal(cqMap)
-	if err != nil {
-		return nil, fmt.Errorf("failed to marshal ClusterQueue to YAML: %w", err)
-	}
-
-	return cqBytes, nil
-}
-
-func (g *GKEOrchestrator) buildFlavorResources(coveredMap map[string]bool, fc FlavorCapacity, fname string) []map[string]interface{} {
-	var resources []map[string]interface{}
-	_, hasPathways := g.capacity.Flavors["pathways-flavor"]
-	for res := range coveredMap {
-		var quota = g.getNominalQuota(res, fc, fname)
-		if hasPathways && isTPUFlavor(fname) {
-			switch res {
-			case "cpu":
-				quota = "999999"
-			case "memory":
-				quota = "999999T"
-			}
-		}
-		resources = append(resources, map[string]interface{}{
-			"name":         res,
-			"nominalQuota": quota,
-		})
-	}
-	sort.Slice(resources, func(i, j int) bool {
-		return resources[i]["name"].(string) < resources[j]["name"].(string)
-	})
-	return resources
-}
-
-func buildResourceGroup(coveredMap map[string]bool, flavors []map[string]interface{}) map[string]interface{} {
-	var coveredResources []string
-	for res := range coveredMap {
-		coveredResources = append(coveredResources, res)
-	}
-	sort.Strings(coveredResources)
-	if len(coveredResources) == 0 {
-		return nil
-	}
-	return map[string]interface{}{
-		"coveredResources": coveredResources,
-		"flavors":          flavors,
-	}
-}
-
-func (g *GKEOrchestrator) renderResourceFlavor(name string, nodeLabels map[string]string) ([]byte, error) {
-	rfMap := map[string]interface{}{
-		"apiVersion": "kueue.x-k8s.io/" + kueueAPIVersion,
-		"kind":       "ResourceFlavor",
-		"metadata": map[string]interface{}{
-			"name": name,
-		},
-	}
-	if len(nodeLabels) > 0 {
-		filteredLabels := make(map[string]string)
-		for k, v := range nodeLabels {
-			if k != tpuTopologyLabel &&
-				!strings.HasPrefix(k, "cloud.google.com/gke-tpu-slice-") &&
-				!strings.HasPrefix(k, "cloud.google.com/gke-tpu-partition-") {
-				filteredLabels[k] = v
-			}
-		}
-		if len(filteredLabels) > 0 {
-			rfMap["spec"] = map[string]interface{}{
-				"nodeLabels": filteredLabels,
-			}
-		}
-	}
-	return yaml.Marshal(rfMap)
-}
-
-func (g *GKEOrchestrator) EnsureResourceFlavors() error {
-	logging.Info("Ensuring Kueue ResourceFlavors exist...")
-	for name, fc := range g.capacity.Flavors {
-		logging.Info("Ensuring ResourceFlavor '%s'...", name)
-		rfBytes, err := g.renderResourceFlavor(name, fc.NodeLabels)
-		if err != nil {
-			return fmt.Errorf("failed to render ResourceFlavor %s: %w", name, err)
-		}
-		if err := g.applyManifests(rfBytes, "resource-flavor.yaml"); err != nil {
-			return fmt.Errorf("failed to apply ResourceFlavor %s: %w", name, err)
-		}
-	}
-	return nil
 }
 
 func (g *GKEOrchestrator) installJobSetCRD(jobSetManifestsURL string) error {
@@ -628,6 +95,17 @@ type k8sEndpointSliceList struct {
 	} `json:"items"`
 }
 
+func (eps *k8sEndpointSliceList) HasReadyEndpoint() bool {
+	for _, item := range eps.Items {
+		for _, ep := range item.Endpoints {
+			if ep.Conditions.Ready && len(ep.Addresses) > 0 {
+				return true
+			}
+		}
+	}
+	return false
+}
+
 func (g *GKEOrchestrator) waitForJobSetWebhook() error {
 	logging.Info("Waiting for JobSet webhook service to be ready...")
 	res := g.executor.ExecuteCommand("kubectl", "rollout", "status", "deployment/jobset-controller-manager", "-n", "jobset-system", "--timeout=600s")
@@ -640,15 +118,12 @@ func (g *GKEOrchestrator) waitForJobSetWebhook() error {
 		cmdEndpoints := g.executor.ExecuteCommand("kubectl", "get", "endpointslice", "-l", "kubernetes.io/service-name=jobset-webhook-service", "-n", "jobset-system", "-o", "json")
 		if cmdEndpoints.ExitCode == 0 {
 			var eps k8sEndpointSliceList
-			if err := json.Unmarshal([]byte(cmdEndpoints.Stdout), &eps); err == nil {
-				for _, item := range eps.Items {
-					for _, ep := range item.Endpoints {
-						if ep.Conditions.Ready && len(ep.Addresses) > 0 {
-							logging.Info("JobSet webhook service endpoints are available.")
-							return nil
-						}
-					}
-				}
+			if err := json.Unmarshal([]byte(cmdEndpoints.Stdout), &eps); err != nil {
+				return fmt.Errorf("failed to unmarshal jobset endpointslice json: %w", err)
+			}
+			if eps.HasReadyEndpoint() {
+				logging.Info("JobSet webhook service endpoints are available.")
+				return nil
 			}
 		}
 		time.Sleep(3 * time.Second)
@@ -657,185 +132,30 @@ func (g *GKEOrchestrator) waitForJobSetWebhook() error {
 	return fmt.Errorf("timed out waiting for jobset-webhook-service endpoints to be available")
 }
 
-func parseVersion(v string) (int, int, int) {
-	v = strings.TrimPrefix(v, "v")
-	parts := strings.Split(v, ".")
-	major, _ := strconv.Atoi(parts[0])
-	minor := 0
-	patch := 0
-	if len(parts) > 1 {
-		minor, _ = strconv.Atoi(parts[1])
-	}
-	if len(parts) > 2 {
-		patchStr := parts[2]
-		// Strip pre-release and build metadata
-		patchStr = strings.Split(patchStr, "-")[0]
-		patchStr = strings.Split(patchStr, "+")[0]
-		patch, _ = strconv.Atoi(patchStr)
-	}
-	return major, minor, patch
-}
-
-func (g *GKEOrchestrator) waitForKueueWebhook() error {
-	res := g.executor.ExecuteCommand("kubectl", "rollout", "status", "deployment/kueue-controller-manager", "-n", "kueue-system", "--timeout=600s")
-	if res.ExitCode != 0 {
-		podDetails := g.getKueuePodDetails()
-		return fmt.Errorf("kueue controller manager failed to become ready: %s\n%s%s", res.Stderr, res.Stdout, podDetails)
-	}
-
-	version, err := g.GetKueueVersion()
-	if err != nil {
-		logging.Warn("Failed to get Kueue version, defaulting to Endpoints check: %v", err)
-		version = defaultKueueVersion // Fallback to older version behavior
-	}
-
-	major, minor, _ := parseVersion(version)
-	useEndpointSlice := major > 0 || (major == 0 && minor > 13)
-
-	endpointsReady := false
-	for i := 0; i < 40; i++ {
-		ready, err := g.checkKueueEndpoints(useEndpointSlice)
-		if err != nil {
-			return err
-		}
-		if ready {
-			logging.Info("Kueue webhook service endpoints are available.")
-			endpointsReady = true
-			break
-		}
-		time.Sleep(3 * time.Second)
-	}
-
-	if !endpointsReady {
-		return fmt.Errorf("timed out waiting for kueue-webhook-service endpoints to be available")
-	}
-
-	// Active probe to ensure webhook is processing requests
-	logging.Info("Probing Kueue webhook readiness...")
-	probeManifest := `apiVersion: kueue.x-k8s.io/` + kueueAPIVersion + `
-kind: ResourceFlavor
-metadata:
-  name: gcluster-webhook-probe
-`
-	f, err := os.CreateTemp("", "gcluster-webhook-probe-*.yaml")
-	if err != nil {
-		return fmt.Errorf("failed to create probe manifest file: %w", err)
-	}
-	probeFile := f.Name()
-	if _, err := f.Write([]byte(probeManifest)); err != nil {
-		f.Close()
-		return fmt.Errorf("failed to write probe manifest: %w", err)
-	}
-	f.Close()
-	defer os.Remove(probeFile)
-
-	for i := 0; i < 20; i++ {
-		res := g.executor.ExecuteCommand("kubectl", "apply", "-f", probeFile)
-		if res.ExitCode == 0 {
-			logging.Info("Kueue webhook is fully operational.")
-			g.executor.ExecuteCommand("kubectl", "delete", "-f", probeFile, "--ignore-not-found")
-			return nil
-		}
-		time.Sleep(5 * time.Second)
-	}
-
-	return fmt.Errorf("timed out waiting for Kueue webhook to become operational")
-}
-
-func (g *GKEOrchestrator) getKueuePodDetails() string {
-	podRes := g.executor.ExecuteCommand("kubectl", "get", "pods", "-n", "kueue-system", "-l", "control-plane=controller-manager", "-o", "json")
-	var podDetails string
-	if podRes.ExitCode == 0 {
-		var podList struct {
-			Items []struct {
-				Metadata struct {
-					Name string `json:"name"`
-				} `json:"metadata"`
-				Status struct {
-					ContainerStatuses []struct {
-						Name  string `json:"name"`
-						State struct {
-							Waiting struct {
-								Reason  string `json:"reason"`
-								Message string `json:"message"`
-							} `json:"waiting"`
-						} `json:"state"`
-					} `json:"containerStatuses"`
-				} `json:"status"`
-			} `json:"items"`
-		}
-		if err := json.Unmarshal([]byte(podRes.Stdout), &podList); err == nil {
-			for _, item := range podList.Items {
-				for _, cs := range item.Status.ContainerStatuses {
-					if cs.State.Waiting.Reason != "" {
-						podDetails += fmt.Sprintf("\n  - Pod %s: %s (%s)", item.Metadata.Name, cs.State.Waiting.Reason, cs.State.Waiting.Message)
-					}
-				}
-			}
-		}
-	}
-	return podDetails
-}
-
-func (g *GKEOrchestrator) checkKueueEndpoints(useEndpointSlice bool) (bool, error) {
-	var cmdEndpoints shell.CommandResult
-	if useEndpointSlice {
-		cmdEndpoints = g.executor.ExecuteCommand("kubectl", "get", "endpointslice", "-l", "kubernetes.io/service-name=kueue-webhook-service", "-n", "kueue-system", "-o", "json")
-	} else {
-		cmdEndpoints = g.executor.ExecuteCommand("kubectl", "get", "endpoints", "kueue-webhook-service", "-n", "kueue-system", "-o", "json")
-	}
-
-	if cmdEndpoints.ExitCode != 0 {
-		return false, nil
-	}
-
-	if useEndpointSlice {
-		var eps k8sEndpointSliceList
-		if err := json.Unmarshal([]byte(cmdEndpoints.Stdout), &eps); err == nil {
-			for _, item := range eps.Items {
-				for _, ep := range item.Endpoints {
-					if ep.Conditions.Ready && len(ep.Addresses) > 0 {
-						return true, nil
-					}
-				}
-			}
-		}
-	} else {
-		var eps struct {
-			Subsets []struct {
-				Addresses []struct {
-					Ip string `json:"ip"`
-				} `json:"addresses"`
-			} `json:"subsets"`
-		}
-		if err := json.Unmarshal([]byte(cmdEndpoints.Stdout), &eps); err == nil {
-			for _, subset := range eps.Subsets {
-				if len(subset.Addresses) > 0 {
-					return true, nil
-				}
-			}
-		}
-	}
-	return false, nil
-}
-
 func (g *GKEOrchestrator) isJobSetCRDInstalled() (bool, error) {
 	res := g.executor.ExecuteCommand("kubectl", "get", "crd", "jobsets.jobset.x-k8s.io")
 	if res.ExitCode == 0 {
 		return true, nil
 	}
-	if strings.Contains(res.Stderr, "not found") || strings.Contains(res.Stdout, "NotFound") {
+	errStr := strings.ToLower(res.Stderr + " " + res.Stdout)
+	if strings.Contains(errStr, "not found") || strings.Contains(errStr, "notfound") {
 		logging.Info("JobSet CRD not found.")
 		return false, nil
+	}
+	if strings.Contains(errStr, "forbidden") {
+		logging.Warn("Insufficient RBAC permissions to read JobSet CRD (403 Forbidden). Assuming JobSet is installed in shared cluster.")
+		return true, nil
 	}
 	return false, fmt.Errorf("failed to check for JobSet CRD: %s\n%s", res.Stderr, res.Stdout)
 }
 
 func (g *GKEOrchestrator) getHTTPClient() HTTPClient {
-	if g.httpClient != nil {
-		return g.httpClient
-	}
-	return &http.Client{Timeout: 30 * time.Second}
+	g.httpOnce.Do(func() {
+		if g.httpClient == nil {
+			g.httpClient = &http.Client{Timeout: 30 * time.Second}
+		}
+	})
+	return g.httpClient
 }
 
 func (g *GKEOrchestrator) downloadManifests(url string) ([]byte, error) {
@@ -1088,41 +408,4 @@ func (g *GKEOrchestrator) checkDynamicSlicingViaGKE() bool {
 		}
 	}
 	return false
-}
-
-func (g *GKEOrchestrator) checkKueueInstallPermissions(version string) error {
-	logging.Info("Verifying cluster permissions for Kueue installation...")
-	checks := []struct {
-		verb     string
-		resource string
-	}{
-		{"create", "namespaces"},
-		{"create", "customresourcedefinitions.apiextensions.k8s.io"},
-		{"create", "clusterroles.rbac.authorization.k8s.io"},
-		{"create", "clusterrolebindings.rbac.authorization.k8s.io"},
-	}
-
-	for _, c := range checks {
-		res := g.executor.ExecuteCommand("kubectl", "auth", "can-i", c.verb, c.resource)
-		if res.ExitCode != 0 || strings.TrimSpace(res.Stdout) != "yes" {
-			return fmt.Errorf("unable to re-install kueue to version %s, this could be a permission issue. Please contact your cluster administrator for updating KUEUE settings", version)
-		}
-	}
-	return nil
-}
-
-func (g *GKEOrchestrator) validatePriorityClass(requestedPriority string) error {
-	if requestedPriority == "" {
-		return nil
-	}
-
-	existing, err := g.getClusterPriorityClasses()
-	if err != nil {
-		return err
-	}
-
-	if !slices.Contains(existing, requestedPriority) {
-		return fmt.Errorf("priority class %q does not exist in the cluster. Available priority classes are: %v", requestedPriority, existing)
-	}
-	return nil
 }
